Polling for interrupt
This gets rid of the need for an interrupt by polling for midnight. It
will only send a message when no one is pumping. It also should get rid
of getTimeHour() and instead just reference it to the external RTCC. The
timeStampValue has also been modified so it references the external
RTCC. NONE OF THESE HAVE BEEN TESTED.