Finished GitHub Guide
This contains how to / information guides on:
a. Signing up for GitHub
b. Downloading GitHub and saving a new document
c. Opening a current document in GitHub
d. How to make a folder in GitHub
e. Transferring who is in charge of a repository.
f. Making a repository
g. Vocabulary