ADC issues are a hardware problem
When this code is run on the v07r01 board it works as expected, but run
on a breadboard it just idles the UART line. This will be tested once I
get batteries for the v07r01 or r02 boards or if I happen to solve the
hardware issue. My suspicion is that the caps are too far from the
board, but that's hard to fix with a breadboard.