Working Final Code with test bits in it
This code appears to work. The problem had been that the
waterPresenceSensorOnOffPin hadn't been made into an output, and it
needed to be.
Problems:
1) The volume is not accurate, appears too large.
2) Fona is not turning off after a text message is sent
3) ? SendTextMessage takes a long time because we added delays in it.
4) ? Need to double check RTCC's accuracy

Debug stuff included:
This texts every hour when the handle is not moving.
There are also unused pins that are being toggled and read with the
logic analyzer (pin 4, pin 5, pin 7, and pin 9)
Pin 4: toggles when handle isn't moving.
Pin 5: toggles in priming loop
Pin 7: toggles in volume calculation loop.
Pin 9: toggles in Leak Rate Loop.