Ghana changes volume loop problem
This code used Putty over the tx and rx line so doesn't send a message
and has many sendMessage() notifications through out.
The volume loop has a problem.
Before this change the volume loop would stay here for too long of a
time. We'd like it to leave if someone isn't moving it up OR DOWN.
Currently it does not.