Depth Sensor and Angle Adjustment
Depth Sensor Code was added but is UNTESTED. Since the Depth Sensor does
not go on every unit, you must let the board know if the depth sensor is
attached or not by making depthSensorInUse a 1 if it is used, or
depthSensorInUse a 0 if it is not in use.

getHandleAngle() was adjusted but is NOT working since it does not have
all of the constants that it needs.