Creation of EEPROM files
This is code written by KWHR. I did modify it by including #ifndef and
#define. (I'm currently not extremely confident on header files, and
believed you needed to include those. I will look more into this).

Email that Tom Austin sent to me when he sent me the files:
Hi Jacqui,

I figured out how to read/write the EEPROM last year in Mahadaga and
then handed it off to Nathan.
Attached in file EEPROM_14-06-03.h is what looks like what I handed him.
It is fairly well commented.

I also attached the code from the latest program that went to TCZ so you
can see what we currently have.

IIRC, there were two items that made it very hard to figure out.

1
When you reprogram the PIC the EEPROM is affected - especially if you
have a 'default' value such as:
int __attribute__ ((space(eedata))) EEpassword0 = '2';
Each time you reprogram the PIC the value will be reset making it near
impossible to test if the EEPROM was updated with a different value.
Any checking of an updated EEPROM value must be done without
reprogramming the PIC.
While simple, this hung me up for several hours when figuring things
out.

2
The EEPROM functions have to have hard coded addressing. You will notice
that the EEPROM code is not efficient and is copied and copied -
variables are not passed into a function for addressing - it is all
*hard coded*. The problem lies in maintaining and passing along the
memory location - it just did not work. Hence explicitly copying code
for each EEPROM item to be read / written.

There may be ways around each of these items, but they are what hung me
up the most when first getting it to work.

My suggestion - understand the EEPROM is a very simple and easy
environment.
My test set up was on a breadboard and simply blinked an LED the number
of times contained in the EEPROM with a method to increase the number.
Once you get it working and understand it enough, then implement it into
your core code.
Once you get it it is fairly simple, but until then it can be a beast to
get working like you expect.

This is slowly coming back to me, so if you have any more questions, I
might be able to answer better later on.

Good Luck!

Tom