Testing Accelerometer Code
Putty is currently disabled, because I was checking the new board to at
least make sure that the Fona would respond. It did so that's great!

The updated Accelerometer code in getHandleAngle did not seem to be
working, it kept reading 180 degrees when there was no "cut off" and
20degrees when there was a cut off.

Future: Plan to take a look at the math and the code and make sure that
the math is correct and that the code matches the correct math.